refactor: change file structure, delete unused fonts
