+++ src/UI/buyNowBtn/buyNowBtn.scss
