+++ src/UI/buyNowBtn/buyNowBtn.tsx
