+++ src/UI/carousel/carousel.scss
