+++ src/UI/dualSlider/dualSlider.scss
