+++ src/UI/dualSlider/dualSlider.tsx
