+++ src/components/filterPanel/filterPanel.tsx
@@ -3,7 +3,7 @@ import { IProduct } from '../../types';
 import { CategoryItem } from '../categoryItem/categoryItem';
 import { useSearchParams } from 'react-router-dom';
 import { useState } from 'react';
-import { DualSlider } from '../dualSlider/dualSlider';
+import { DualSlider } from '../../UI/dualSlider/dualSlider';
 import { products } from '../../data/data';
 import {
   getMaxPriceFilteredProducts,
@@ -14,7 +14,7 @@ import {
   getMinStock,
   getMaxStockFilteredProducts,
   getMinStockFilteredProducts,
-} from '../utility/utility';
+} from '../../utility/utility';
 
 interface IFilterPanelProps {
   filteredProducts: IProduct[];
