+++ src/index.scss
@@ -1,28 +1,3 @@
-@font-face {
-  font-family: "Ablation"; 
-  src: url("./fonts/Ablation\ ExtraBold.eot") format('eot'), 
-  url("./fonts/Ablation\ ExtraBold.woff") format('woff'),
-  url("./fonts/Ablation\ ExtraBold.ttf") format('ttf'); 
-  font-style: normal; 
-  font-weight: normal; 
-} 
-@font-face {
-  font-family: "YanoneKaffeesatz"; 
-  src: url("./fonts/YanoneKaffeesatz-Regular.eot") format('eot'), 
-  url("./fonts/YanoneKaffeesatz-Regular.woff") format('woff'),
-  url("./fonts/YanoneKaffeesatz-Regular.ttf") format('ttf'); 
-  font-style: normal; 
-  font-weight: normal; 
-} 
-@font-face {
-  font-family: "YanoneKaffeesatz"; 
-  src: url("./fonts/YanoneKaffeesatz-Medium.eot") format('eot'), 
-  url("./fonts/YanoneKaffeesatz-Medium.woff") format('woff'),
-  url("./fonts/YanoneKaffeesatz-Medium.ttf") format('ttf'); 
-  font-style: normal; 
-  font-weight: bold; 
-} 
-
 body {
   color: #694430;
   font-weight: 500;
