+++ src/pages/productItemPage/productItemPage.scss
