+++ src/pages/productsPage/productsPage.scss
